Replace param $persist with $batch_args on filter process
Replace param $persist with $batch_args on filter process.
The persist value still exists within the $response associative array,
which is the primary value being filtered.

While $index is duplicated (it appears both in the $response as well as
a separate parameter), I don't think this is necessary for $persist (it
is an edge case use).

Since $batch_args aren't part of the response, I think it makes sense to
make this a separate parameter.

This is a breaking change for anyone using the parameter $persist that
was formerly being passed through on the filter call.

See #1